requirements.txt
Flask==0.10.1
Flask-Login==0.2.4
Flask-OAuth==0.12
Jinja2==2.7
MarkupSafe==0.18
WTForms==1.0.4
Werkzeug==0.9.1
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.1
httplib2==0.8
itsdangerous==0.21
mongoengine==0.8.2
oauth2==1.5.211
pymongo==2.5.2
wsgiref==0.1.2
